Changed number types from int to double

+++ particle.h
@@ -1,45 +1,47 @@
 #ifndef _PARTICLE_H_
 #define _PARTICLE_H_
 
+#include <cmath>  // for the unit vector calculation
+
 // declaration of a class representing every particle
 // e.g. pollen grain, dust particle, H2O molecule or O2 molecule etc.
 class Particle
 {
 private:
     // radius of the particle
-    int r;
+    double r;
     // mass of the particle
-    int m;
+    double m;
     // speed of the particle
-    float v;
+    double v;
     // position of the particle
-    int x;
-    int y;
+    double x;
+    double y;
     // unit vector of the particle - representation of the movement direction
-    int vx;
-    int vy;
+    double vx;
+    double vy;
 public:
     Particle();
-    Particle(int radius, int mass, int velocity, int x_pos, int y_pos, int vector_x, int vector_y);
+    Particle(double radius, double mass, double velocity, double x_pos, double y_pos, double vector_x, double vector_y);
     
     // working with the position
-    int getX() const { return x; }
-    int getY() const { return y; }
-    void setX(int new_x) { x = new_x; }
-    void setY(int new_y) { y = new_y; }
+    double getX() const { return x; }
+    double getY() const { return y; }
+    void setX(double new_x) { x = new_x; }
+    void setY(double new_y) { y = new_y; }
 
     // working with the vector
-    int getVx() const { return vx; }
-    int getVy() const { return vy; }
-    void setVx(int new_vx) { vx = new_vx; }
-    void setVy(int new_vy) { vy = new_vy; }
+    double getVx() const { return vx; }
+    double getVy() const { return vy; }
+    void setVx(double new_vx) { vx = new_vx; }
+    void setVy(double new_vy) { vy = new_vy; }
 
     // working with the velocity
-    float getV() const { return v; }
-    void setV(int new_v) { v = new_v; }
+    double getV() const { return v; }
+    void setV(double new_v) { v = new_v; }
 
-    int radius() const { return r; }  // returns the radius of the particle
-    int mass() const { return m; }  // returns the mass of the particle
+    double radius() const { return r; }  // returns the radius of the particle
+    double mass() const { return m; }  // returns the mass of the particle
 
     ~Particle();
 };
